# BLD: Coalesce binops to cut `_sparsetools.so` library size in half