#### Reference issue
No issue, just noticed this opportunity while working on another related project.

#### What does this implement/fix?
CSR sparse matrix element-wise binary operators are implemented as templates. The binop operation is fairly large and instantiated many times, so binops account for a significant fraction of both compilation time and library size.

Binops are largely memory bound. That means slightly increasing the cost of the op, even in the inner loop, has little visible performance impact. Therefore the less common operators can be grouped into one implementation with the operand chosen at runtime. This allows two template instantiations to cover all operands. Two because the comparators (greater, less, not_equal_to, etc) return bool while the other operators return T.

Plus and Minus are left alone.

### Build Impact

`_sparsetools.so` goes from 4.0 MB to 2.2 MB.

Total `python build.py dev` runtime drops by about 5%, but the very large`csr.cxx.o` and `bsr.cxx.o` compilation time drops significantly (big red and purple blocks on the bottom rows):

main branch:
![main timeline](https://github.com/scipy/scipy/assets/2730364/88124bf1-6870-4487-bf5e-7ed439edf76f)

this PR:
![smallcsr timeline](https://github.com/scipy/scipy/assets/2730364/3d2e5e2f-d10b-4993-9e0a-66569535f3c6)

These may be the dominating steps for folks with 20+ core machines so those folks may see a noticeable speedup.

### Runtime Impact

I'm seeing up to 5% impact on elementwise multiply on very large (1M+ nnz) matrices, and less on smaller ones. Note that #19765 speeds up these methods by more than this amount.