@rgommers 2023-12-28
Nice idea, thanks Adam!

> I'm seeing up to 5% impact on elementwise multiply on very large (1M+ nnz) matrices, and less on smaller ones. Note that #19765 speeds up these methods by more than this amount.

It looks like the impact is larger on some benchmarks. It still sounds like a good idea to me though, given that the sum of this PR and gh-19765 halves binary size _and_ improves performance. 

Of course benchmarking results are always tricky to get right. Here what I did was remove the `--quick` flag from `cmd_compare` in `dev.py` (needed to get stable results in this case) and then run
```
python dev.py bench --compare main -t sparse.Arithmetic
```

With the default matrix sizes in the benchmarks (`poisson2d(250)`) ~= 250,000 elements:

```
| Change   | Before [bee9b9ae] <main>   | After [b667240a] <smallcsr>   |   Ratio | Benchmark (Parameter)                                      |
|----------|----------------------------|-------------------------------|---------|------------------------------------------------------------|
| +        | 3.18±0.04ms                | 4.30±0.3ms                    |    1.35 | sparse.Arithmetic.time_arithmetic('dia', 'BB', 'multiply') |
| +        | 3.17±0.05ms                | 4.12±0.2ms                    |    1.3  | sparse.Arithmetic.time_arithmetic('csc', 'BB', 'multiply') |
| +        | 3.24±0.1ms                 | 4.11±0.1ms                    |    1.27 | sparse.Arithmetic.time_arithmetic('coo', 'BB', 'multiply') |
| +        | 3.16±0.02ms                | 3.96±0.1ms                    |    1.26 | sparse.Arithmetic.time_arithmetic('csr', 'BB', 'multiply') |
| +        | 1.32±0.02ms                | 1.64±0.01ms                   |    1.24 | sparse.Arithmetic.time_arithmetic('csc', 'AA', 'multiply') |
| +        | 9.57±0.1ms                 | 11.8±2ms                      |    1.23 | sparse.Arithmetic.time_arithmetic('dia', 'AB', '__add__')  |
| +        | 3.02±0.07ms                | 3.58±0.06ms                   |    1.18 | sparse.Arithmetic.time_arithmetic('csc', 'AB', 'multiply') |
| +        | 1.21±0.02ms                | 1.39±0.02ms                   |    1.15 | sparse.Arithmetic.time_arithmetic('csr', 'AA', 'multiply') |
| +        | 9.54±0.1ms                 | 10.8±0.3ms                    |    1.13 | sparse.Arithmetic.time_arithmetic('dia', 'BA', 'multiply') |
| +        | 3.09±0.02ms                | 3.40±0.1ms                    |    1.1  | sparse.Arithmetic.time_arithmetic('csc', 'BA', 'multiply') |
| +        | 2.99±0.03ms                | 3.30±0.03ms                   |    1.1  | sparse.Arithmetic.time_arithmetic('csr', 'BA', 'multiply') |
| +        | 6.92±0.04ms                | 7.50±0.4ms                    |    1.08 | sparse.Arithmetic.time_arithmetic('coo', 'AB', '__sub__')  |
| +        | 7.03±0.2ms                 | 7.43±0.2ms                    |    1.06 | sparse.Arithmetic.time_arithmetic('coo', 'BA', 'multiply') |
| +        | 9.30±0.2ms                 | 9.86±0.2ms                    |    1.06 | sparse.Arithmetic.time_arithmetic('dia', 'BA', '__add__')  |
| -        | 986±20μs                   | 929±7μs                       |    0.94 | sparse.Arithmetic.time_arithmetic('csr', 'AA', '__sub__')  |
```

Increasing the sizes to (`poisson2d(1000)`) ~= 4M elements (this takes a while to run, comment out all but `multiply` to make it run faster):
```
| Change   | Before [bee9b9ae] <main>   | After [b667240a] <smallcsr>   |   Ratio | Benchmark (Parameter)                                      |
|----------|----------------------------|-------------------------------|---------|------------------------------------------------------------|
| +        | 62.5±3ms                   | 78.2±2ms                      |    1.25 | sparse.Arithmetic.time_arithmetic('csr', 'BB', 'multiply') |
| +        | 72.7±1ms                   | 85.3±1ms                      |    1.17 | sparse.Arithmetic.time_arithmetic('csc', 'BB', 'multiply') |
| +        | 68.4±0.6ms                 | 80.3±0.3ms                    |    1.17 | sparse.Arithmetic.time_arithmetic('dia', 'BB', 'multiply') |
| +        | 72.8±1ms                   | 83.6±1ms                      |    1.15 | sparse.Arithmetic.time_arithmetic('coo', 'BB', 'multiply') |
| +        | 29.2±0.9ms                 | 33.5±0.5ms                    |    1.14 | sparse.Arithmetic.time_arithmetic('csc', 'AA', 'multiply') |
| +        | 24.0±0.3ms                 | 27.3±0.5ms                    |    1.14 | sparse.Arithmetic.time_arithmetic('csr', 'AA', 'multiply') |
| +        | 65.6±1ms                   | 72.1±0.4ms                    |    1.1  | sparse.Arithmetic.time_arithmetic('csc', 'AB', 'multiply') |
| +        | 57.1±2ms                   | 62.7±2ms                      |    1.1  | sparse.Arithmetic.time_arithmetic('csr', 'BA', 'multiply') |
| +        | 65.9±0.3ms                 | 70.4±0.3ms                    |    1.07 | sparse.Arithmetic.time_arithmetic('csc', 'BA', 'multiply') |
| -        | 15.6±0.3ms                 | 14.6±0.2ms                    |    0.94 | sparse.Arithmetic.time_arithmetic('csr', 'AA', '__sub__')  |
```
So it looks like ~20% impact on average for element-wise multiplication here. This is with the default dev build settings on a x86-64 Linux machine; I didn't check whether release settings (`-O3` vs `-O2`, no debug info) make a difference here.