@perimosocordiae 2024-03-06
Thanks for the PR, @alugowski! This is a useful experiment, and I really appreciate the analysis here.

My current position is that it seems like the performance hit may not be worth the reduction in build size, but I would love to get more benchmark data if folks want to try this out on their local environments. There may also be a middle-ground option that does less runtime dispatch but still reduces the total number of template instantiations.