@tylerjereddy 2024-03-15
There's still a bit of debate about the tradeoffs here, and it certainly isn't crucial before I branch for `1.13.0` to support NumPy 2, so I'll just bump the milestone.