@tylerjereddy 2024-05-27
Bumping milestone for same reason noted above--the tradeoffs are still debatable but the idea is considered interesting.